pyproject.toml
[project]
name = "rag-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "black>=26.3.1",
    "chromadb>=1.5.7",
    "duckduckgo-search>=8.1.1",
    "flake8>=7.3.0",
    "huggingface-hub>=1.10.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "langsmith[claude-agent-sdk]>=0.7.30",
    "loguru>=0.7.3",
    "lxml>=6.0.3",
    "pandas>=3.0.2",
    "pre-commit>=4.5.1",
    "pymupdf4llm>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.0",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
